+++ src/models/region.rs
@@ -562,6 +562,8 @@ mod tests {
             "list.txt".into(),
             "local".into(),
             "".into(),
+            0,
+            0,
         );
         let mut r = Region::new(
             "r".into(),
@@ -635,6 +637,8 @@ mod tests {
             "".into(),
             "local".into(),
             "".into(),
+            0,
+            0,
         );
         let parent = joined(
             "parent",
@@ -671,6 +675,8 @@ mod tests {
             "".into(),
             "local".into(),
             "".into(),
+            0,
+            0,
         );
         let r2 = Region::new(
             "bc".into(),